📣 - Android live ride announement (#272)
closes #271 

depends on #256

---------
