+++ app/i18n/he.json
@@ -159,7 +159,8 @@
     "announcement": {
       "title": "בטר רייל לייב",
       "subtitle": "הנסיעה ברכבת הולכת להשתנות עם",
-      "description": "קבלו התראות על איחורי רכבות, זמן הגעה, הוראות נסיעה ועוד - והכל על גבי מסך הנעילה של המכשיר.",
+      "description": "התראות על איחורי רכבות, זמן הגעה, הוראות נסיעה ועוד - והכל על גבי מסך הנעילה של המכשיר.",
+      "androidDescription": "התראות על איחורי רכבות, זמן הגעה, הוראות נסיעה ועוד.",
       "weMadeAGuide": "הכנו מדריך קצר על השימוש בפיצ׳ר שישנה את הדרך בה אתם נוסעים ברכבת."
     },
     "startRide": {
@@ -170,7 +171,9 @@
     },
     "liveActivity": {
       "title": "לייב אקטיביטי",
+      "androidTitle": "נסיעה פעילה",
       "description": "אחרי שהתחלתם נסיעה, היא תתווסף למסך הנעילה של המכשיר ותתעדכן אוטומטית עם איחורים והתקדמות הרכבת.",
+      "androidDescription": "אחרי שהתחלתם נסיעה, היא תתווסף למסך הנוטיפקציות ותתעדכן אוטומטית עם איחורים והתקדמות הרכבת.",
       "tip": "טיפ: אפשר להתחיל את הנסיעה עד שעה לפני שהרכבת יוצאת מהתחנה. ככה תוכלו לדעת אם יש צורך לרוץ על מנת לתפוס את הרכבת, או שאפשר להשאר רגועים."
     },
     "dynamicIsland": {
@@ -179,7 +182,7 @@
       "description2": "נתוני הנסיעה יופיעו על גבי האי הדינאמי, אליו תוכלו לגשת מכל מקום במכשיר."
     },
     "supportUs": {
-      "title": "היי, אנחנו גיא ומתן",
+      "title": "היי! אנחנו גיא ומתן",
       "description1": "אנחנו שני נוסעי רכבת עם אהבה גדולה לפיתוח אפליקציות.",
       "description3": "העבודה על בטר רייל כרוכה במאות שעות עבודה, ואנחנו משלמים מכיסינו הפרטי על מנת לממן את פיתוח האפליקצייה.",
       "description4": "אם בטר רייל משפרת לכם את חווית הנסיעה ברכבת, אנא תמכו בנו.",
