+++ app/screens/live-announcement/activity-announcement-screen.ios.tsx
