+++ app/screens/live-announcement/activity-announcement-screen.ts
@@ -1,0 +1,9 @@
+import { Platform } from "react-native"
+
+import { ActivityAnnouncementScreen as ActivityAnnouncementScreenIOS } from "./activity-announcement-screen.ios"
+import { ActivityAnnouncementScreen as ActivityAnnouncementScreenAndroid } from "./activity-announcement-screen.android"
+
+export const ActivityAnnouncementScreen = Platform.select({
+  ios: ActivityAnnouncementScreenIOS,
+  android: ActivityAnnouncementScreenAndroid,
+})
