+++ app/screens/live-announcement/index.ts
@@ -1,5 +1,5 @@
 export * from "./activity-announcement-screen"
-export * from "./start-ride-announcement-screen"
 export * from "./live-announcement-screen"
+export * from "./start-ride-screen"
 export * from "./dynamic-island-screen"
 export * from "./support-us-announcement-screen"
