+++ app/screens/live-announcement/start-ride-screen.tsx
@@ -1,0 +1,8 @@
+import { Platform } from "react-native"
+import { StartRideAnnouncement as StartRideAnnouncementAndroid } from "./start-ride-announcement-screen.android"
+import { StartRideAnnouncement as StartRideAnnouncementIOS } from "./start-ride-announcement-screen.ios"
+
+export const StartRideAnnouncement = Platform.select({
+  ios: StartRideAnnouncementIOS,
+  android: StartRideAnnouncementAndroid,
+})
